requirements.txt
fastapi
uvicorn
python-telegram-bot
google-generativeai
web3
python-dotenv
pydantic
requests
aiohttp
httpx
pytest
black
flake8
mypy